refactor: implement DataEventSource and DataEventSourceEncoder for Server-Sent Events
